New variable CMAKE_REAL_INSTALL_PREFIX
The template files for shell profiles where filled using the
CMAKE_INSTALL_PREFIX. However, when building the RPM,
CMAKE_INSTALL_PREFIX is set with the RPM BUILDROOT which is different
of the real target installation when the RPM is deployed. The solution
is to set a new variable set CMAKE_REAL_INSTALL_PREFIX which sets PATH
in the profile scripts with the correct path.